// src/app/shared/models/basket.ts
export interface Product {
  id: string;
  goodName: string;
  goodPageUrl: string;
  brandName: string;
  price: number;
  discount: number;
  finalPrice: number;
  count: number;
  imageUrl: string;
  goodType: number;
}
export interface Good {
  id: string;
  goodParent: IGoodParentDto | null;
  brand: IBrandDto | null;
  subBrand: IBrandDto | null;
  pageUrl: string;
  pageTitle: string;
  pageDescription: string;
  technicalCode: string;
  code: string;
  description: string;
  status: number;
  statusName: string;
  commentStatus: boolean;
  reviewUrl: string;
  previewUrl: string;
  previewSummeryUrl: string;
  goodImages: IGoodImageDto[];
  goodTags: string[] | null;
  goodQuantityParameters: string[];
  goodStorageAddresses: IGoodStorageAddressDto[];
  prices: IPrice[];
  userGroupPrice: IUserGroupPriceDto | null;
  stock: number;
  storageIODetails: IStorageIODetail[];
  persianLastBuyDate: string;
  lastBuyPrice: number;
  lastSellPrice: number;
  persianLastSalePriceUpdateDate: string,
  specialCode: string,
  isFocus?: boolean;
  buyLimit: number | null;
}
export interface IGoodGroupTypeDto {
  id: number;
  typeName: string;
  code: string;
  searched?: boolean
}
export interface IGoodGroupDto {
  id: number;
  name: string;
  nameEn: string | null;
  goodGroupTypeDto: IGoodGroupTypeDto;
  code: string;
  parent: IGoodGroupDto | null
  description: string | null;
  discount: number;
  status: number;
  statusTitle?: string;
  index: number;
  pageUrl: string | null;
  pageTitle: string | null;
  pageDescription: string | null;
  groupImageDtos?: IGroupImageDto[];
  inverseParents?: string[];
  goodGroupQualityParameterDtos?: string[];
  goodGroupQuantityParameterDtos?: string[];
  goodGoodGroupId?: string,
  searched?: boolean;
  children?: IGoodGroupDto[];
}
export interface IGroupImageDto {
  imageUrl: string;
  addressType: number;
}
export interface IGoodParentDto {
  id: string;
  name: string;
  unit: IGoodUnitDto | null;
  code: string;
  goodGroups: IGoodGroupDto[];
  barcodeSize: number;
  goodParentTags: IGoodParentTag[];
  price?: number;
  quantity?: number;
  total?: number;
}
export interface IGoodUnitDto {
  id: number;
  name: string;
}
export interface IBrandDto {
  id: number;
  name: string;
  nameEn: string | null;
  code: string;
  image: string;
  description: string;
  pageUrl: string;
  pageTitle: string;
  pageDescription: string;
  term: ITermDto | null;
  termId?: number;
}
export interface ITermDto {
  id: number;
  code: string;
  termDetails: ITermDetailDto[]
}
export interface ITermDetailDto {
  id: number;
  value: string
}
export interface IGoodImageDto {
  id: string;
  imageUrl: string;
  type: number;
}
export interface IGoodStorageAddressDto {
  storageSectionName: string;
  row: string;
  column: string;
  shelf: string;
}
export interface IGoodUnitDto {
  id: number;
  name: string;
}
export interface IPrice {
  userGroup: string;
  userGroupId: number;
  defaultPrice: number;
  userPrice: number;
  discount: number;
  discountPercent: number;
  finalPrice: number;
}
export interface IUserGroupPriceDto {
  id: number;
  userGroupId: number | null;
  userGroupName: string | null;
  goodId: string;
  price: number;
  discount: number;
}
export interface IStorageIODetail {
  storageIO: IStorageIO;
  count: number;
}
export interface IStorageIO {
  storage: IStorage;
  storageOperation: IInvoiceAndPayType;
}
// ...
export interface IGoodParentTag {
  id: string;
  name: string;
}